Improve documentation of register constraints.
While looking at PR target/64532- I realized we haven't documented all
the register constraints. I'm not documenting the other immediate
constraints as it is not clear to me how much of that is actually
useful yet and I don't have the time this afternoon to clean this up.

Built documentation and looked at it.

Applied.

Ramana



git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@219847 138bc75d-0d04-0410-961f-82ee72b054a4