rs6000: Fix PR67344
The "*and<mode>3_imm_dot_shifted" pattern is a define_insn_and_split,
like most "dot" patterns: if its output is not assigned cr0 but some
other cr reg, it splits to a non-dot insn and a compare.

Unfortunately that non-dot insn will clobber cr0 as well.  We could
add another clobber (with "=X,x"), but then that second alternative
is never useful; instead, just remove that second alternative.


2015-08-25  Segher Boessenkool  <segher@kernel.crashing.org>

	PR target/67344
	* config/rs6000/rs6000.md (*and<mode>3_imm_dot_shifted): Change to
	a define_insn, remove second alternative.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@227182 138bc75d-0d04-0410-961f-82ee72b054a4